+++ org.eclipse.jgit/src/org/eclipse/jgit/transport/TransportSftp.java
@@ -58,9 +58,9 @@ import java.util.Map;
 import java.util.Set;
 import java.util.TreeMap;
 
-import org.eclipse.jgit.JGitText;
 import org.eclipse.jgit.errors.NotSupportedException;
 import org.eclipse.jgit.errors.TransportException;
+import org.eclipse.jgit.internal.JGitText;
 import org.eclipse.jgit.lib.Constants;
 import org.eclipse.jgit.lib.ObjectId;
 import org.eclipse.jgit.lib.ObjectIdRef;
@@ -242,13 +242,14 @@ public class TransportSftp extends SshTransport implements WalkTransport {
 					if (!files.containsKey(in))
 						continue;
 
-					mtimes.put(n, ent.getAttrs().getMTime());
+					mtimes.put(n, Integer.valueOf(ent.getAttrs().getMTime()));
 					packs.add(n);
 				}
 
 				Collections.sort(packs, new Comparator<String>() {
 					public int compare(final String o1, final String o2) {
-						return mtimes.get(o2) - mtimes.get(o1);
+						return mtimes.get(o2).intValue()
+								- mtimes.get(o1).intValue();
 					}
 				});
 			} catch (SftpException je) {
